requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
diskcache==5.6.3
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
idna==3.6
InstructorEmbedding==1.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-experimental==0.0.49
langchainplus-sdk==0.0.20
langsmith==0.0.83
llama_cpp_python==0.2.36
MarkupSafe==2.1.4
marshmallow==3.20.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numexpr==2.9.0
numpy==1.26.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.2
pillow==10.2.0
pydantic==1.10.14
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.22.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==11.0.2
yarl==1.9.4
